Move NTIA check to ntia_checker.py
