+++ ntia_conformance_checker/__init__.py
@@ -1,3 +1,6 @@
 """Export functions for usage as library."""
 
+__all__ = ["NTIAChecker", "SbomChecker"]
+
+from ntia_conformance_checker.ntia_checker import NTIAChecker
 from ntia_conformance_checker.sbom_checker import SbomChecker
